migrate classes to separate typescript files
