+++ src/TagStatus.ts
@@ -1,0 +1,10 @@
+
+enum TagStatus {
+    pending = '🕑',
+    successful = '✅',
+    caution = '⚠️',
+    failed = '❌',
+    global = '🌍'
+}
+
+export { TagStatus }
